+++ Assets/Source/Script/EntriesEndsHandler.cs
@@ -4,7 +4,7 @@ namespace Balloondle.Script
 {
     public abstract class EntriesEndsHandler : ScriptableObject
     {
-        public abstract void OnEntryStart(ulong entryId);
-        public abstract void OnEntryEnd(ulong entryId);
+        public abstract void OnEntryStart(int entryId);
+        public abstract void OnEntryEnd(int entryId);
     }
 }
